Separating DocsGenerator to another file
